Tests moved to separate folders for clarity. HomepageFlowTest added.
