+++ tests/Application/CreateChallengeControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests;
+namespace App\Tests\Application;
 
 use App\Entity\Exercise;
 use App\Entity\MuscleGroup;
