+++ tests/Behaviour/HomepageFlowTest.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace App\Tests;
+namespace App\Tests\Behaviour;
 
 use Exception;
 use Facebook\WebDriver\Exception\NoSuchElementException;
 use Facebook\WebDriver\Exception\TimeoutException;
 use Symfony\Component\Panther\PantherTestCase;
 
-class FlowTest extends PantherTestCase
+class HomepageFlowTest extends PantherTestCase
 {
     /**
      * @throws NoSuchElementException
@@ -40,10 +40,10 @@ class FlowTest extends PantherTestCase
      * @throws NoSuchElementException
      * @throws TimeoutException
      */
-    public function testUserClicksAround(): void
+    public function testGuestInteractsWithTheHomepage(): void
     {
         $client = static::createPantherClient();
-        $client->request('GET', '/user/register');
+        $client->request('GET', '/');
 
         $client->waitFor('nav');
         $client->clickLink('Login');
@@ -54,12 +54,17 @@ class FlowTest extends PantherTestCase
         $client->clickLink('Register');
         $client->waitFor('h1');
         $this->assertSelectorTextContains('h1', 'REGISTER');
+
+        $client->waitFor('nav');
+        $client->clickLink('Home');
+        $client->waitFor('h1');
+        $this->assertSelectorTextContains('h1', 'Welcome traveler, perhaps you struggle with whatever in your life... And you think my web-site is late.');
     }
 
     /**
      * @throws Exception
      */
-    public function testUserCanLogin(): void
+    public function testLoggedInUserInteractsWithHomepage(): void
     {
         $client = static::createPantherClient();
         $credentials = $this->generateUser();
@@ -73,55 +78,9 @@ class FlowTest extends PantherTestCase
         $client->submit($form);
 
         $this->assertSelectorExists('.welcome-message');
-    }
-
-    /**
-     * @throws NoSuchElementException
-     * @throws TimeoutException
-     * @throws Exception
-     */
-    public function testUserStartsANewQuestCompletesItAndLogsOut(): void
-    {
-        $client = static::createPantherClient();
-        $credentials = $this->generateUser();
-
-        $crawler = $client->request('GET', '/user/login');
-
-        $form = $crawler->selectButton('Sign in')->form([
-            '_username' => $credentials['email'],
-            '_password' => $credentials['password'],
-        ]);
-        $client->submit($form);
-
-        $this->assertSelectorExists('.welcome-message');
-        $client->waitFor('.welcome-message h1');
-
-        $client->clickLink('Begin Questing');
-
+        $client->waitFor('nav');
+        $client->clickLink('Home');
         $client->waitFor('h1');
-        $this->assertSelectorTextContains(
-            'h1',
-            'You have not completed a single quest? I know a man in white robes, that would be disappointed'
-        );
-
-        $client->clickLink('Start your first quest');
-
-        $client->waitFor('ul');
-        $this->assertPageTitleContains('Quest Started!');
-
-        $challengeForm = $crawler->selectButton('MARK AS DONE')->form([
-
-        ]);
-        $client->submit($challengeForm);
-        $client->waitFor('.welcome-message');
-        $this->assertSelectorTextContains('h2', 'List of completed quests, Gondorian:');
-        $client->clickLink('View');
-        $client->waitFor('ul');
-        $client->clickLink('Back');
-        $client->waitFor('.welcome-message');
-
-        $client->clickLink('Logout');
-        $client->waitFor('h1');
-        $this->assertSelectorTextContains('h1', 'Please sign in');
+        $this->assertSelectorTextContains('h1', 'Welcome traveler, perhaps you struggle with whatever in your life... And you think my web-site is late.');
     }
 }
