+++ tests/Unit/DashboardControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests;
+namespace App\Tests\Unit;
 
 use App\Controller\DashboardController;
 use App\Repository\ChallengeRepository;
