+++ tests/Unit/RandomExerciseFetcherTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests;
+namespace App\Tests\Unit;
 
 use App\Entity\Exercise;
 use App\Exception\NotEnoughExercisesException;
